Plugins: move PanelPluginMeta to grafana/ui (#16804)
