+++ public/app/features/dashboard/components/DashExportModal/DashboardExporter.ts
@@ -4,7 +4,7 @@ import config from 'app/core/config';
 import { DashboardModel } from '../../state/DashboardModel';
 import DatasourceSrv from 'app/features/plugins/datasource_srv';
 import { PanelModel } from 'app/features/dashboard/state';
-import { PanelPluginMeta } from 'app/types/plugins';
+import { PanelPluginMeta } from '@grafana/ui';
 
 interface Input {
   name: string;
