+++ public/test/specs/helpers.ts
@@ -3,8 +3,7 @@ import config from 'app/core/config';
 import * as dateMath from 'app/core/utils/datemath';
 import { angularMocks, sinon } from '../lib/common';
 import { PanelModel } from 'app/features/dashboard/state/PanelModel';
-import { PanelPluginMeta } from 'app/types';
-import { RawTimeRange } from '@grafana/ui/src/types';
+import { PanelPluginMeta, RawTimeRange } from '@grafana/ui';
 
 export function ControllerTestContext(this: any) {
   const self = this;
